Support convenient separation of objects
Create objects from files with a shared key, then use that key in the
master template to delegate to partials. This makes specifying objects
low overhead and keeps related objects' descriptions to only a few
files: a map, their construction (which mostly specifies a texture),
and the relevant interaction text file, and any state transitions
required.