Move save file into a single local storage key
This was needed cleanup anyway, but also important since LocalStorage
is shared for the entire domain, meaning the game clutters the entire
namespace for the domains hosting it. The only downside is a less clean
debugging interface to the save file.

In the process also clean storage of state properties on the
StateMachine. The actual values are now held in _props, but proxy
getters are installed for known properties: this remains compatible with
the rest of the code, which assumes all properties are accessible by
field access, including computed properties defined as getters, and also
means it is no longer possible to assign any known property directly
(that is, without using @set).